Add jetty-plugin config suitable for multi-module hot deployment
Change jetty-maven-plugin config that will be used in project
generated from this archetype so that jetty will monitor for
changes also in the actual ville-exercise-project. Actual ville-
exercise-project is sibling of the test-stub-project so to get
changes done to it notified by jetty, add that project
as extraClassPath-entry and scanTarget-entry.