Use Vaadin- and vexer-versions from the vexer-project in vexer-archetype
Changed vexer-archetype in such a way that current Vaadin and vexer versions
in use in vexer-archetype project are included in the archetype when
the archetype jar is generated. The user does not have to type in those
versions at archetype:create-time anymore. Also added dependency
to commons.codec 1.8 to generated stub's pom, to ensure that the version
needed by ville-standardutils is used.